% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_info.R
\name{get_eea_ddb_extension_info}
\alias{get_eea_ddb_extension_info}
\alias{get_eea_ddb_vars}
\alias{get_eea_version}
\title{Access information about the EEA Air Quality Download Service API}
\usage{
get_eea_ddb_extension_info()

get_eea_ddb_vars()

get_eea_version()
}
\description{
These functions aren't immediately useful for downloading air quality data,
but provide assorted information about the EEA AQ database.
}
\author{
Jack Davison
}
