% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_eea_measurements.R
\name{download_eea_parquet_files}
\alias{download_eea_parquet_files}
\alias{download_eea_parquet_async}
\alias{download_eea_parquet_urls}
\alias{download_eea_summary}
\title{Import measurements from the EEA Air Quality Download Service API}
\usage{
download_eea_parquet_files(
  countries = "ES",
  cities = "Madrid",
  pollutants = NULL,
  datetime_start = Sys.Date() - 30,
  datetime_end = Sys.Date(),
  dataset = 1L,
  aggregation_type = "hour",
  file = tempfile(fileext = ".zip")
)

download_eea_parquet_async(
  countries = "ES",
  cities = "Madrid",
  pollutants = NULL,
  datetime_start = Sys.Date() - 30,
  datetime_end = Sys.Date(),
  dataset = 1L,
  aggregation_type = "hour"
)

download_eea_parquet_urls(
  countries = "ES",
  cities = "Madrid",
  pollutants = NULL,
  datetime_start = Sys.Date() - 30,
  datetime_end = Sys.Date(),
  dataset = 1L,
  aggregation_type = "hour"
)

download_eea_summary(
  countries = "ES",
  cities = "Madrid",
  pollutants = NULL,
  datetime_start = Sys.Date() - 30,
  datetime_end = Sys.Date(),
  dataset = 1L,
  aggregation_type = "hour"
)
}
\arguments{
\item{countries}{A vector of country codes from \code{\link[=import_eea_countries]{import_eea_countries()}}. If
\code{NULL}, data from all countries will be imported, which is not recommended
given the final file size.}

\item{cities}{A vector of cities in the given \code{countries} from
\code{\link[=import_eea_cities]{import_eea_cities()}}. If \code{NULL}, data from all cities in the given
\code{countries} will be imported.}

\item{pollutants}{A vector of pollutant notations or IDs from
\code{\link[=import_eea_pollutants]{import_eea_pollutants()}}. If \code{NULL}, data for all pollutants will be
imported.}

\item{datetime_start, datetime_end}{Start and end date times, provided as
\code{POSIXct} or \code{Date} R objects.}

\item{dataset}{The value of the dataset. One of:
\enumerate{
\item Unverified data transmitted continuously (Up-To-Date/UTD/E2a) data from
the beginning of 2023.
\item Verified data (E1a) from 2013 to 2022 reported by countries by 30
September each year for the previous year.
\item Historical Airbase data delivered between 2002 and 2012 before Air
Quality Directive 2008/50/EC entered into force
}}

\item{aggregation_type}{represents whether the data collected is obtaining
the values:
\enumerate{
\item Hourly data.
\item Daily data.
\item Variable intervals (different than the previous observations such as
weekly, monthly, etc.)
}}

\item{file}{The file to write the zip file to; should ideally end in
\code{".zip"}. Defaults to a temporary file using \code{\link[=tempfile]{tempfile()}}.}
}
\value{
One of:
\itemize{
\item \code{\link[=download_eea_parquet_files]{download_eea_parquet_files()}}: the \code{file} argument - the path to the downloaded ZIP file.
\item \code{\link[=download_eea_parquet_async]{download_eea_parquet_async()}}: a path to the URL at which the ZIP file will be made available.
\item \code{\link[=download_eea_parquet_urls]{download_eea_parquet_urls()}}: a character vector of URLS to each parquet file.
\item \code{\link[=download_eea_summary]{download_eea_summary()}}: a numeric list, with names \code{numberFiles} and \code{size}.
}
}
\description{
These functions are wrappers for the 'ParquetFile' endpoints of the European
Environment Agency's Air Quality Download Service API.
\code{\link[=download_eea_parquet_files]{download_eea_parquet_files()}} download zip file with all the filtered
parquet files in it. \code{\link[=download_eea_parquet_async]{download_eea_parquet_async()}} downloads this in the
background; the function returns an URL in which the file to be downloaded
will be generated. \code{\link[=download_eea_parquet_urls]{download_eea_parquet_urls()}} returns a list of URLs
corresponding to the filtered parquets. \code{\link[=download_eea_summary]{download_eea_summary()}} estimates
the number of files and their size, but does not return any data.
}
\author{
Jack Davison
}
