% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_eea_vocabulary.R
\name{import_eea_cities}
\alias{import_eea_cities}
\alias{import_eea_countries}
\alias{import_eea_pollutants}
\title{Import metadata files from the EEA Air Quality Download Service}
\usage{
import_eea_cities(country)

import_eea_countries()

import_eea_pollutants()
}
\arguments{
\item{country}{A vector of country codes from \code{\link[=import_eea_countries]{import_eea_countries()}}. Only
used in \code{\link[=import_eea_cities]{import_eea_cities()}}.}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
This collection of functions import information on available countries,
cities and pollutants, formatting them as \link[tibble:tibble-package]{tibbles}.
}
\author{
Jack Davison
}
