% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eea_raster.R
\name{get_eea_raster_files}
\alias{get_eea_raster_files}
\alias{get_eea_raster_async}
\title{Import raster data from the EEA Air Quality Download Service API}
\usage{
get_eea_raster_files(
  country = NULL,
  city = NULL,
  pollutants = NULL,
  datetime_start = Sys.Date() - 10,
  datetime_end = Sys.Date(),
  email = NULL,
  output_format = "zip",
  file = tempfile(fileext = paste0(".", output_format))
)

get_eea_raster_async(
  country = NULL,
  city = NULL,
  pollutants = NULL,
  datetime_start = Sys.Date() - 10,
  datetime_end = Sys.Date(),
  email = NULL,
  output_format = "zip",
  file = tempfile(fileext = paste0(".", output_format))
)
}
\arguments{
\item{country}{A single country code from \code{\link[=import_eea_countries]{import_eea_countries()}}.}

\item{city}{A single city in the given \code{country} from \code{\link[=import_eea_cities]{import_eea_cities()}}.
If \code{NULL}, data from all cities in the given \code{country} will be imported.}

\item{pollutants}{A vector of pollutant IDs from \code{\link[=import_eea_pollutants]{import_eea_pollutants()}}
(e.g., \code{"pm10"}).}

\item{datetime_start, datetime_end}{Start and end date times, provided as
\code{POSIXct}, \code{Date} or \code{integer} R objects. If an \code{integer} is provided, this
should represent the year of interest; for \code{datetime_start} this will be
represent the first hour of the year and for \code{datetime_end} it will
represent the last hour of the year, meaning providing the same integer to
each will return a year of data.}

\item{email}{Optional field to identify the user who make the download and
improve the communication if problems are detected.}

\item{output_format}{One of \code{"zip"} or \code{"tiff"}.}

\item{file}{The file to write the zip file to; should ideally end in
\code{".zip"}. Defaults to a temporary file using \code{\link[=tempfile]{tempfile()}}.}
}
\value{
One of:
\itemize{
\item \code{\link[=get_eea_raster_files]{get_eea_raster_files()}}: the \code{file} argument - the path to the downloaded ZIP file.
\item \code{\link[=get_eea_raster_async]{get_eea_raster_async()}}: a path to the URL at which the ZIP file will be made available.
}
}
\description{
These functions are wrappers for the 'Raster' endpoints of the European
Environment Agency's Air Quality Download Service API.
\itemize{
\item \code{\link[=get_eea_raster_files]{get_eea_raster_files()}} downloads a zip file with all the filtered parquet
files in it.
\item \code{\link[=get_eea_raster_async]{get_eea_raster_async()}} prepares this in the background; the
function returns an URL in which the file to be downloaded will be generated.
}
}
\author{
Jack Davison
}
