% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eea_version.R
\name{get_eea_version}
\alias{get_eea_version}
\title{Access information about the EEA Air Quality Download Service API}
\usage{
get_eea_version()
}
\description{
These functions aren't immediately useful for downloading air quality data,
but provide assorted information about the EEA AQ database.
}
\author{
Jack Davison
}
