% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_eea_raster.R
\name{import_eea_raster}
\alias{import_eea_raster}
\title{Conveniently import European Air Quality CAMS data into R}
\usage{
import_eea_raster(
  country = NULL,
  city = NULL,
  pollutants = NULL,
  datetime_start = Sys.Date() - 10,
  datetime_end = Sys.Date(),
  ...,
  .endpoint = "files"
)
}
\arguments{
\item{country}{A single country code from \code{\link[=import_eea_countries]{import_eea_countries()}}.}

\item{city}{A single city in the given \code{country} from \code{\link[=import_eea_cities]{import_eea_cities()}}.
If \code{NULL}, data from all cities in the given \code{country} will be imported.}

\item{pollutants}{A vector of pollutant IDs from \code{\link[=import_eea_pollutants]{import_eea_pollutants()}}
(e.g., \code{"pm10"}).}

\item{datetime_start, datetime_end}{Start and end date times, provided as
\code{POSIXct}, \code{Date} or \code{integer} R objects. If an \code{integer} is provided, this
should represent the year of interest; for \code{datetime_start} this will be
represent the first hour of the year and for \code{datetime_end} it will
represent the last hour of the year, meaning providing the same integer to
each will return a year of data.}

\item{...}{Not currently used.}

\item{.endpoint}{Provided for consistency with \code{\link[=import_eea_monitoring]{import_eea_monitoring()}}.
Currently only \code{"files"} is supported, which uses \code{\link[=get_eea_raster_files]{get_eea_raster_files()}}.}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
This function is a convenient way to use EEA AQ modelling data in an R
session. It calls \code{\link[=get_eea_raster_files]{get_eea_raster_files()}} to download the .tiff files into a
temporary directory and reads them into the R session using \code{\link[terra:rast]{terra::rast()}}.
}
\author{
Jack Davison
}
