% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_eea_stations.R
\name{import_eea_stations}
\alias{import_eea_stations}
\title{Import Air Quality Site Metadata from Dataflow D}
\usage{
import_eea_stations(countries = NULL)
}
\arguments{
\item{countries}{A vector of country codes from \code{\link[=import_eea_countries]{import_eea_countries()}}. If
\code{NULL}, data from all countries will be imported.}
}
\description{
Imports data from the \href{https://discomap.eea.europa.eu/App/AQViewer/index.html?fqn=Airquality_Dissem.b2g.measurements}{data flow D data viewer}.
This is useful for binding onto datasets obtained using
\code{\link[=get_eea_parquet_files]{get_eea_parquet_files()}} to assign monitoring data with relevant site
information. The user must select a subset of countries using the \code{countries} argument.
}
\author{
Jack Davison
}
