% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_eea_monitoring.R
\name{import_eea_monitoring}
\alias{import_eea_monitoring}
\title{Conveniently import European Air Quality data into R}
\usage{
import_eea_monitoring(
  countries = "ES",
  cities = "Madrid",
  pollutants = NULL,
  dataset = 1L
)
}
\arguments{
\item{countries}{A vector of country codes from \code{\link[=import_eea_countries]{import_eea_countries()}}. If
\code{NULL}, data from all countries will be imported, which is not recommended
given the final file size.}

\item{cities}{A vector of cities in the given \code{countries} from
\code{\link[=import_eea_cities]{import_eea_cities()}}. If \code{NULL}, data from all cities in the given
\code{countries} will be imported.}

\item{pollutants}{A vector of pollutant notations or IDs from
\code{\link[=import_eea_pollutants]{import_eea_pollutants()}}. If \code{NULL}, data for all pollutants will be
imported.}

\item{dataset}{The value of the dataset. One of:
\enumerate{
\item Unverified data transmitted continuously (Up-To-Date/UTD/E2a) data from
the beginning of the year.
\item Verified data (E1a) from 2013, reported by countries by 30 September
each year for the previous year.
\item Historical Airbase data delivered between 2002 and 2012 before Air
Quality Directive 2008/50/EC entered into force
}}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
This function is a convenient way to use EEA AQ monitoring data in an R
session. It calls \code{\link[=download_eea_parquet_urls]{download_eea_parquet_urls()}}, reads each file using
\code{\link[arrow:read_parquet]{arrow::read_parquet()}}, removes unnecessary columns, and merges useful
columns from \code{\link[=import_eea_stations]{import_eea_stations()}}. As this function uses
\code{\link[=download_eea_parquet_urls]{download_eea_parquet_urls()}}, users are not able to specify a date range or
aggregation type.
}
\author{
Jack Davison
}
