% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_eea_stations.R
\name{import_eea_stations}
\alias{import_eea_stations}
\title{Import Air Quality Site Metadata from Dataflow D}
\usage{
import_eea_stations(..., .cache = TRUE)
}
\arguments{
\item{...}{Not currently used; reserved for future functionality.}

\item{.cache}{This function works by installing a CSV to a temporary
directory and reading it into R. When \code{.cache} is \code{TRUE}, this only occurs
once per R session. When \code{.cache} is \code{FALSE}, this CSV is re-downloaded
every time the function is called.}
}
\description{
Imports data from the \href{https://discomap.eea.europa.eu/App/AQViewer/index.html?fqn=Airquality_Dissem.b2g.measurements}{data flow D data viewer}.
This is useful for binding onto datasets obtained using
\code{\link[=download_eea_parquet_files]{download_eea_parquet_files()}} to assign monitoring data with relevant site
information.
}
\author{
Jack Davison
}
